Demonstrate Mail service
Implements a Mail service that sends a dummy message using the MS Graph API. Test the mail service by making a GET request to `/testSendMail`.

Note this service is not paramaterized currently, so it sends the same message to the same person (our pal Bruce). We'll add params later, but just
want to get it working in a more basic form first.

Does not implement a mock Mail service. Need to do this as well before all is said and done.

The service "works", but fails due to an MS/Azure permissions issue.