Add a test for meetings matcher utility
Microsoft refers to the same meeting by different ids. This makes it
impossible to tell if two meetings are the same. Impossible, that is,
unless we write our own meetings matcher. Which we did.

This commit adds a test for the meetingsMatch function. Hopefully it
provides some measure of documentation for the reasoning behind this
strange little utility.