Add a service that adds a new user
This functionality depends upon a "service user", which acts on behalf of the Bookit app. The service user must be set up, and our app must have the correct permissions set in Azure. These setup steps are documented in the README and the docs folder.

The service adds a contact to the service user's contact list.

The service user is currently stored in
`src/config/identities/index.ts`. This is not the best way to do this,
but it works for the moment.