Add service to fetch internal (i.e. Designit) users
The rest service now serves all users (Internal and External) at /users.

We're normalizing the two types of users to a common schema. This has
some consequences. For instance, external users are given empty strings
for names, and internal users have empty strings for role.

Here is a known weak link: The MS User endpoint gives us rooms as well
as human users. We filter out the rooms in the internal user service,
but this filter relies on a naming convention. This is brittle -- it
would break as soon as someone decides on another naming convention for
rooms.