Respond with valid JSON when user service errors
Currently, we are passing any errors with the user service back to the
client. This breaks the client because the client is expecting valid
JSON.

This commit sends an empty array when the user service throws an error.
It logs the error on the server side, but does not expose the error to
the client.